remove type.hpp from string.hpp
