+++ modules/runtime/include/containers/string.hpp
@@ -75,17 +75,6 @@ struct argument_formatter<skr_guid_t>
 }
 }
 
-#include "type/type_id.hpp"
-
-namespace skr {
-namespace type {
-// {214ed643-54bd-4213-be37-e336a77fde84}
-SKR_RTTI_INLINE_REGISTER_BASE_TYPE(skr::string, 0x214ed643, 0x54bd, 0x4213, 0xbe, 0x37, 0xe3, 0x36, 0xa7, 0x7f, 0xde, 0x84, "214ED643-54BD-4213-BE37-E336A77FDE84");
-// {b799ba81-6009-405d-9131-e4b6101660dc}
-SKR_RTTI_INLINE_REGISTER_BASE_TYPE(skr::string_view, 0xb799ba81, 0x6009, 0x405d, 0x91, 0x31, 0xe4, 0xb6, 0x10, 0x16, 0x60, 0xdc, "B799BA81-6009-405D-9131-E4B6101660DC");
-} // namespace type
-} // namespace skr
-
 #include "serde/binary/blob_fwd.h"
 namespace skr::binary
 {
