+++ modules/runtime/include/containers/vector.hpp
@@ -6,56 +6,8 @@ namespace skr
 using eastl::vector;
 }
 
-#include "type/type.hpp"
-
-namespace skr
-{
-namespace type
-{
-struct DynArrayStorage
-{
-    uint8_t* begin;
-    uint8_t* end;
-    uint8_t* capacity;
-};
-// skr::vector<T>
-struct DynArrayType : skr_type_t {
-    const struct skr_type_t* elementType;
-    skr::string name;
-    RUNTIME_API uint64_t Num(void* data) const;
-    RUNTIME_API void* Get(void* data, uint64_t index) const;
-    RUNTIME_API void Reset(void* data, uint64_t size) const;
-    RUNTIME_API void Reserve(void* data, uint64_t size) const;
-    RUNTIME_API void Resize(void* data, uint64_t size) const;
-    RUNTIME_API void* Insert(void* data, uint64_t index) const;
-    RUNTIME_API void Erase(void* data, uint64_t index, bool bKeepOrder) const;
-    DynArrayType(const skr_type_t* elementType)
-        : skr_type_t{ SKR_TYPE_CATEGORY_DYNARR }
-        , elementType(elementType)
-    {
-    }
-private:
-    void Grow(void* data, uint64_t size) const;
-};
-RUNTIME_API const skr_type_t* make_dynarray_type(const skr_type_t* type);
-template <class V, class T>
-struct type_of_vector {
-    static const skr_type_t* get()
-    {
-        return make_dynarray_type(type_of<T>::get());
-    }
-};
-
-template <class T, class Allocator>\
-struct type_of<skr::vector<T, Allocator>> : type_of_vector<skr::vector<T, Allocator>, T> 
-{
-};
-
-}
-}
-
+#include "serde/binary/serde.h"
 // binary reader
-#include "misc/traits.hpp"
 #include "serde/binary/reader_fwd.h"
 
 namespace skr
@@ -113,7 +65,7 @@ struct SerdeCompleteChecker<binary::ReadTrait<skr::vector<V, Allocator>>>
 }
 
 // binary writer
-#include "serde/binary/reader_fwd.h"
+#include "serde/binary/writer_fwd.h"
 
 namespace skr
 {
@@ -211,6 +163,6 @@ struct VectorWriterBitpacked
 
 template <class V, class Allocator>
 struct SerdeCompleteChecker<binary::WriteTrait<const skr::vector<V, Allocator>&>>
-    : std::bool_constant<is_complete_serde_v<json::WriteTrait<V>>> {
+    : std::bool_constant<is_complete_serde_v<binary::WriteTrait<V>>> {
 };
 } // namespace skr
