+++ tests/platform/marl-test/marl_test.h
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
+#pragma once
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
