+++ tools/tool_core/xmake.lua
@@ -1,5 +1,6 @@
 shared_module("SkrToolCore", "TOOL_CORE", engine_version)
     set_group("02.tools")
+    add_rules("c++.unity_build", {batchsize = default_unity_batch_size})
     add_files("src/**.cpp")
     public_dependency("SkrRT", engine_version)
     add_includedirs("include", {public = true})
